refactor(kubernetes): Push fields down from ManifestBasedModel
Most of ManifestBasedModel is lookign up fields on a Manifest;
we want to stop using the manifest for all of these lookups because
in most cases we already store the parsed data as part of the CacheData
item (and in the cases we don't, we should start doing that).

In order to simplify moving to that model, push down all the manifest
logic to the implementing classes. It's a bit more code now, but should
make it much easier to understand. We also remove some methods that aren't
necessary because they were added to the superclass even though they are
only needed by one or two implementations.

Add a KubernetesResource interface that keeps track of fields we're
expecting to find on all kubernetes resources (mapping to the correpsonding
interface in deck). This is just helpful to make sure that we get a
compile-time error if we accidentally remove one of these fields.
